requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
pyserial==3.5
numpy==1.24.3
python-dotenv==1.0.0
eventlet==0.33.3
